Updated dependency to Rails Admin 0.6.6 (Bootstrap 3)
