+++ rails_admin_import.gemspec
@@ -5,8 +5,14 @@ require "rails_admin_import/version"
 Gem::Specification.new do |s|
   s.name = "rails_admin_import"
   s.version = RailsAdminImport::VERSION
-  s.authors = ["Steph Skardal"]
-  s.email = "steph@endpoint.com"
+  s.authors = ["Steph Skardal", "Julien Vanier"]
+  s.email = ["steph@endpoint.com", "jvanier@gmail.com"]
+  s.summary = "Import functionality for Rails Admin"
+  s.license     = 'MIT'
+
   s.files = Dir["{app,config,lib}/**/*"] + ["MIT-LICENSE", "Rakefile", "README.md", "RELEASE_NOTES"]
-  s.summary = "Import functionality for rails admin"
+
+  s.add_dependency 'rails', '>= 3.2'
+  s.add_dependency 'rails_admin', '>= 0.6.6'
+  
 end
